services/launchpad-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
pyjwt==2.8.0
prometheus-client==0.19.0
python-multipart==0.0.6